Added Style to Guides and Quote Section
Added a box shadow as a 3D divider to help separate the Episode Guide section from the Quotes section. In the Episode Guide, I added and styled an hr line for more character. It revealed the last child (box), 2019 Revival, is slightly taller than the other 3. The issue seems to stem from the image, even though it shows the same dimensions. Will need to debug further.